requirements.txt
streamlit>=1.36.0
numpy
plotly
requests
pillow
urllib3
typing-extensions
pandas
watchdog
